+++ packages/google/src/google-generative-ai-settings.ts
@@ -20,8 +20,21 @@ export type GoogleGenerativeAIModelId =
   | 'gemini-2.0-flash-thinking-exp-01-21'
   | 'gemini-2.0-flash-exp'
   | 'gemini-exp-1206'
+  | 'gemma-3-27b-it'
   | 'learnlm-1.5-pro-experimental'
   | (string & {});
+
+export interface DynamicRetrievalConfig {
+  /**
+   * The mode of the predictor to be used in dynamic retrieval.
+   */
+  mode?: 'MODE_UNSPECIFIED' | 'MODE_DYNAMIC';
+  /**
+   * The threshold to be used in dynamic retrieval. If not set, a system default
+   * value is used.
+   */
+  dynamicThreshold?: number;
+}
 
 export interface GoogleGenerativeAISettings {
   /**
@@ -73,6 +86,15 @@ Optional. When enabled, the model will use Google search to ground the response.
 @see https://cloud.google.com/vertex-ai/generative-ai/docs/grounding/overview
    */
   useSearchGrounding?: boolean;
+
+  /**
+Optional. Specifies the dynamic retrieval configuration.
+
+@note Dynamic retrieval is only compatible with Gemini 1.5 Flash.
+
+@see https://cloud.google.com/vertex-ai/generative-ai/docs/multimodal/ground-with-google-search#dynamic-retrieval
+   */
+  dynamicRetrievalConfig?: DynamicRetrievalConfig;
 }
 
 export interface InternalGoogleGenerativeAISettings