+++ packages/openai/src/openai-chat-language-model.ts
@@ -133,15 +133,15 @@ export class OpenAIChatLanguageModel implements LanguageModelV1 {
       });
     }
 
-    if (
-      getSystemMessageMode(this.modelId) === 'remove' &&
-      prompt.some(message => message.role === 'system')
-    ) {
-      warnings.push({
-        type: 'other',
-        message: 'system messages are removed for this model',
-      });
-    }
+    const { messages, warnings: messageWarnings } = convertToOpenAIChatMessages(
+      {
+        prompt,
+        useLegacyFunctionCalling,
+        systemMessageMode: getSystemMessageMode(this.modelId),
+      },
+    );
+
+    warnings.push(...messageWarnings);
 
     const baseArgs = {
       // model id:
@@ -158,10 +158,10 @@ export class OpenAIChatLanguageModel implements LanguageModelV1 {
         typeof this.settings.logprobs === 'number'
           ? this.settings.logprobs
           : typeof this.settings.logprobs === 'boolean'
-          ? this.settings.logprobs
-            ? 0
-            : undefined
-          : undefined,
+            ? this.settings.logprobs
+              ? 0
+              : undefined
+            : undefined,
       user: this.settings.user,
       parallel_tool_calls: this.settings.parallelToolCalls,
 
@@ -199,11 +199,7 @@ export class OpenAIChatLanguageModel implements LanguageModelV1 {
         this.settings.reasoningEffort,
 
       // messages:
-      messages: convertToOpenAIChatMessages({
-        prompt,
-        useLegacyFunctionCalling,
-        systemMessageMode: getSystemMessageMode(this.modelId),
-      }),
+      messages,
     };
 
     if (isReasoningModel(this.modelId)) {