+++ packages/openai/src/openai-config.ts
@@ -5,4 +5,5 @@ export type OpenAIConfig = {
   url: (options: { modelId: string; path: string }) => string;
   headers: () => Record<string, string | undefined>;
   fetch?: FetchFunction;
+  generateId?: () => string;
 };