+++ packages/openai/src/openai-prepare-tools.ts
@@ -38,8 +38,7 @@ export function prepareTools({
     parameters: JSONSchema7;
   }[];
   function_call?: { name: string };
-
-  toolWarnings: LanguageModelV1CallWarning[];
+  toolWarnings: Array<LanguageModelV1CallWarning>;
 } {
   // when the tools array is empty, change it to undefined to prevent errors:
   const tools = mode.tools?.length ? mode.tools : undefined;