+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,34 @@
 # ai
 
+## 4.1.58
+
+### Patch Changes
+
+- e9897eb: fix (ai/core): move process access into functions and use globalThis
+
+## 4.1.57
+
+### Patch Changes
+
+- 092fdaa: feat (ai/core): add defaultSettingsMiddleware
+
+## 4.1.56
+
+### Patch Changes
+
+- 80be82b: feat (ai/core): add simulateStreamingMiddleware
+- 8109a24: fix (ai/core): limit node imports to types where possible
+
+## 4.1.55
+
+### Patch Changes
+
+- 1531959: feat (ai/core): add MCP client for using MCP tools
+- Updated dependencies [1531959]
+  - @ai-sdk/provider-utils@2.1.12
+  - @ai-sdk/react@1.1.22
+  - @ai-sdk/ui-utils@1.1.18
+
 ## 4.1.54
 
 ### Patch Changes