+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/vue
 
+## 1.1.21
+
+### Patch Changes
+
+- Updated dependencies [1531959]
+  - @ai-sdk/provider-utils@2.1.12
+  - @ai-sdk/ui-utils@1.1.18
+
 ## 1.1.20
 
 ### Patch Changes